requirements.txt
neat-python~=2.0.0
seaborn~=0.13.2
pandas~=2.3.3
numba~=0.64.0
scipy~=1.15.3
matplotlib~=3.10.8
numpy~=2.2.6
pygame~=2.6.1